# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。",
        "title": "Wavelet API",
        "contact": {
            "name": "Wavelet",
            "url": "https://github.com/Rain-kl/Wavelet"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/admin/status": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "获取后端服务运行状态、Goroutine、内存指标等详细统计数据，需要管理员权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "获取系统状态信息",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Any"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.SystemStatusResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    },
                    "403": {
                        "description": "无管理员权限",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/status/log-database": {
            "get": {
                "security": [
                    {
                        "SessionCookie": []
                    }
                ],
                "description": "返回当前日志主库、迁移状态、各库保留天数与合法迁移目标，需要管理员权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "获取日志数据库状态",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Any"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.LogDatabaseStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    },
                    "403": {
                        "description": "无管理员权限",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.LogDatabaseStatus": {
            "type": "object",
            "properties": {
                "active_database": {
                    "type": "string"
                },
                "available_targets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "migration": {
                    "type": "string"
                },
                "retention_days": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "admin.SystemStatusResponse": {
            "type": "object",
            "properties": {
                "alloc": {
                    "type": "string"
                },
                "buck_hash_sys": {
                    "type": "string"
                },
                "frees": {
                    "type": "integer"
                },
                "gc_sys": {
                    "type": "string"
                },
                "heap_alloc": {
                    "type": "string"
                },
                "heap_idle": {
                    "type": "string"
                },
                "heap_inuse": {
                    "type": "string"
                },
                "heap_objects": {
                    "type": "integer"
                },
                "heap_released": {
                    "type": "string"
                },
                "heap_sys": {
                    "type": "string"
                },
                "last_gc_time": {
                    "type": "string"
                },
                "last_pause": {
                    "type": "string"
                },
                "lookups": {
                    "type": "integer"
                },
                "mallocs": {
                    "type": "integer"
                },
                "mcache_inuse": {
                    "type": "string"
                },
                "mcache_sys": {
                    "type": "string"
                },
                "mspan_inuse": {
                    "type": "string"
                },
                "mspan_sys": {
                    "type": "string"
                },
                "next_gc": {
                    "type": "string"
                },
                "num_gc": {
                    "type": "integer"
                },
                "num_goroutine": {
                    "type": "integer"
                },
                "other_sys": {
                    "type": "string"
                },
                "pause_total_ns": {
                    "type": "string"
                },
                "stack_inuse": {
                    "type": "string"
                },
                "stack_sys": {
                    "type": "string"
                },
                "sys": {
                    "type": "string"
                },
                "total_alloc": {
                    "type": "string"
                },
                "uptime": {
                    "type": "string"
                }
            }
        },
...
        "response.Any": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionCookie": {
            "type": "apiKey",
            "name": "session",
            "in": "cookie"
        }
    }
}